% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organism.R
\name{do_ensembl_organism}
\alias{do_ensembl_organism}
\title{Create a polyApiper organism directory based on ENSEMBL annotation in AnnotationHub}
\usage{
do_ensembl_organism(
  out_path,
  species,
  version,
  hard_extension = 20,
  extension = 2000,
  unstranded_gaps = TRUE,
  omit_biotypes = c(),
  version_dna = version
)
}
\description{
Apologies, you will need to guess the most recent version of your organism on AnnotationHub. The version for the DNA sequence (eg version_dna=100) may be earlier than the version for the annotations (eg version=109).
}
