% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\docType{data}
\name{pipeline_stages}
\alias{pipeline_stages}
\alias{do_pipeline}
\title{Analyse the output of polyApipe.py}
\format{An object of class \code{character} of length 5.}
\usage{
pipeline_stages

do_pipeline(out_path, counts_files, batch_names = "", counts_file_dir,
  peak_info_file, organism, min_peak_count_avg = 0.01,
  min_cell_count_vs_avg = 0.25, remove_mispriming = TRUE,
  utr_or_extension_only = FALSE, do_computeSumFactors = FALSE,
  do_vooma = FALSE, p = 20, stages = pipeline_stages)
}
\description{
Analyse the output of polyApipe.py. The resulting directory can be loaded
with load_banquet.
}
\keyword{datasets}
