% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{do_pipeline_report}
\alias{do_pipeline_report}
\title{Produce a short overview report for the pipeline output}
\usage{
do_pipeline_report(
  out_path,
  input = out_path,
  title = "polyApiper results overview"
)
}
\description{
Produce a short overview report for the pipeline output
}
