% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{enrichment}
\alias{enrichment}
\title{Confident-correlation gene-set enrichment}
\usage{
enrichment(stats, confident_sign, gene_sets, min_size = 2, ad = FALSE)
}
\description{
Effect size is the Pearson correlation between the indicator variable for the gene-set and the statistic.
}
