% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{load_peaks_counts_files_into_sce}
\alias{load_peaks_counts_files_into_sce}
\title{lalalla}
\usage{
load_peaks_counts_files_into_sce(counts_file_list, batch_names,
  peak_info_file, sce_save_path, replace = FALSE, ...)
}
\description{
lalalla
}
