% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{load_peaks_counts_into_sce.NOh5step}
\alias{load_peaks_counts_into_sce.NOh5step}
\title{I think ther was some reason to use DataFrame??}
\usage{
load_peaks_counts_into_sce.NOh5step(counts_file, peak_info_file,
  batch = "", sce_save_path = batch, cell_prefix = batch,
  n_max = -1, min_reads_per_barcode = 100,
  missing_peaks_are_zero = TRUE, ...)
}
\description{
...  Extra options passed to saveHDF5SummarizedExperiment E.g. verbose, replace=TRUE
}
