% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{peaks_shift_weitrix}
\alias{peaks_shift_weitrix}
\title{Shift weitrix from peak counts}
\usage{
peaks_shift_weitrix(peaks_se, organism, n_per_row = 100,
  n_per_col = 100, remove_mispriming = TRUE,
  utr_or_extension_only = TRUE)
}
\description{
Convert a SummarizedExperiment of peak counts into a weitrix of shifts.
}
\details{
Peaks with less than min_present non-zero counts are discarded.

Only genes with two or more peaks are included.

Cells with less than min_present non-missing shifts are discarded.
}
