% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{peaks_weitrices}
\alias{peaks_weitrices}
\title{Shifts and proportions weitrices from peak counts}
\usage{
peaks_weitrices(peaks_se, present_min_count = 1,
  min_peak_present = 0.05, min_cell_present_vs_avg = 0.5,
  remove_mispriming = TRUE, utr_or_extension_only = FALSE,
  do_computeSumFactors = TRUE, do_vooma = FALSE)
}
\description{
Convert a SummarizedExperiment of peak counts into a weitrix of shifts and a weitrix of proportions.
}
\details{
Cells with less than min_cell_count_vs_avg times the average cell count will be discarded.

Peaks with less than an average of min_peak_count_avg counts will be discarded.

Only genes with two or more peaks are included.
}
