% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{read_polyA_peak_file_gtf}
\alias{read_polyA_peak_file_gtf}
\title{Read in the polyA peaks gtf file}
\usage{
read_polyA_peak_file_gtf(gtf_file)
}
\arguments{
\item{peak_info_file}{GTF formatted peak file _specfically_ as output from polyApipe.py}
}
\description{
The polyApipe.py script will write a gtf formatted file with the polyA peak 
locations. This function will read it into a tabular format in R. 
This is not a general gff/gtf reader funtion -  
it will break on probably any other gff!
}
\examples{

peak_info_file <- system.file("extdata", "demo_dataset/demo_polyA_peaks.gff", package = "polyApiper") 
peak_info_table <- read_polyA_peak_file_gtf(peak_info_file)


}
