% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_level.R
\name{se_reassign}
\alias{se_reassign}
\title{Annotate SummarizedExperiment of peak-level data with relationship to genes}
\usage{
se_reassign(se, organism)
}
\description{
Annotate SummarizedExperiment of peak-level data with relationship to genes
}
