% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organism.R
\name{do_ensembl_organism}
\alias{do_ensembl_organism}
\title{Create a polyApiper organism directory based on ENSEMBL annotation in AnnotationHub}
\usage{
do_ensembl_organism(
  out_path,
  species,
  version,
  hard_extension = 20,
  extension = 2000,
  omit_biotypes = c()
)
}
\description{
Create a polyApiper organism directory based on ENSEMBL annotation in AnnotationHub
}
