% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{do_pipeline}
\alias{do_pipeline}
\title{Analyse the output of polyApipe.py}
\usage{
do_pipeline(out_path, counts_files, batch_names = "", counts_file_dir,
  peak_info_file, organism, stages = pipeline_stages)
}
\description{
Analyse the output of polyApipe.py. The resulting directory can be loaded
with load_banquet.
}
