% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{do_pipeline}
\alias{do_pipeline}
\title{Analyse the output of polyApipe.py}
\usage{
do_pipeline(
  out_path,
  counts_file_dir = NULL,
  counts_files = NULL,
  batch_names = "",
  peak_info_file,
  organism,
  cell_name_func = function(batch, cell) paste0(batch, cell),
  cells_to_use = NULL,
  remove_mispriming = TRUE,
  utr_or_extension_only = FALSE,
  peak_min_present = 50,
  peak_min_prop = 0.01,
  do_logNormCounts = TRUE,
  do_computeSumFactors = TRUE,
  title = "polyApiper pipeline run",
  stages = seq_along(PIPELINE_STAGES)
)
}
\arguments{
\item{out_path}{Output directory name.}

\item{counts_file_dir}{A directory containing counts files. Batch names are taken from the basename of the count files without the .tab.gz suffix. Give either this argument or \code{counts_files} but not both.}

\item{counts_files}{Alternative to counts_file_dir. One or more filenames for .tab.gz files produced by polyApipe.py. Give either this argument or \code{counts_file_dir} but not both.}

\item{batch_names}{If using counts_files, give a vector of the batch/sample names, in same order as counts_files}

\item{peak_info_file}{GTF formatted peak file as output from polyApipe.py.}

\item{organism}{Organism directory, as created by \code{do_ensembl_organism}.}

\item{cells_to_use}{Character vector of cells to use.}

\item{remove_mispriming}{Remove peaks considered to be mispriming peaks.}

\item{utr_or_extension_only}{Remove peaks in the 5'UTR, exons, or introns.}

\item{peak_min_present}{A peak is retained if it is present in this number of cells.}

\item{peak_min_prop}{A peak is used for APA and gene expression calculations only if it constitutes at least this proportion of total UMIs for the gene.}

\item{do_logNormCounts}{Compute logcounts for gene and peak expression?}

\item{do_computeSumFactors}{When computing logcounts, use scran::computeSumFactors? If not, unadjusted cell library sizes are used.}

\item{title}{Title to use in report.}

\item{stages}{Stages of the pipeline to run. Stages are listed in PIPELINE_STAGES.}
}
\value{
There is no return value. Results are placed in the \code{out_path} directory. They can be loaded using \code{load_banquet(out_path)}.
}
\description{
Analyse the output of polyApipe.py. The resulting directory can be loaded
with \code{load_banquet}. An HTML report is also produced.
}
\details{
You can choose to run only specific stages of the pipeline with the \code{stages} argument. See the global \code{PIPELINE_STAGES} for possible stages. \code{stages} is a character vector or an integer vector indexing into \code{PIPELINE_STAGES}.
}
