% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\docType{data}
\name{PIPELINE_STAGES}
\alias{PIPELINE_STAGES}
\alias{do_pipeline}
\title{Analyse the output of polyApipe.py}
\format{An object of class \code{character} of length 6.}
\usage{
PIPELINE_STAGES

do_pipeline(out_path, counts_files = NULL, batch_names = "",
  counts_file_dir = NULL, peak_info_file, organism,
  present_min_count = 1, min_peak_present = 0.05,
  min_cell_present_vs_avg = 0.5, remove_mispriming = TRUE,
  utr_or_extension_only = FALSE, do_computeSumFactors = TRUE,
  do_vooma = FALSE, p = 20, title = "polyApiper pipeline run",
  stages = PIPELINE_STAGES)
}
\arguments{
\item{out_path}{Output directory name.}

\item{counts_files}{One or more filenames for .tab.gz files produced by polyApipe.py. Give either this argument or \code{counts_file_dir} but not both.}

\item{batch_names}{A vector of the batch/sample names, in same order as counts_files}

\item{counts_file_dir}{A directory containing counts files. Give either this argument or \code{counts_files} but not both.}

\item{peak_info_file}{GTF formatted peak file as output from polyApipe.py.}

\item{organism}{Organism directory, as created by \code{do_ensembl_organism}.}

\item{stages}{Stages of the pipeline to run.}
}
\description{
Analyse the output of polyApipe.py. The resulting directory can be loaded
with \code{load_banquet}.
}
\details{
You can choose to run only specific stages of the pipeline with the \code{stages} argument, which is a character vector. See the global \code{PIPELINE_STAGES} for possible stages.
}
\keyword{datasets}
