% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banquet.R
\name{load_banquet}
\alias{load_banquet}
\title{Turn a directory into an environment of lazy-loading objects}
\usage{
load_banquet(path)
}
\arguments{
\item{path}{A directory containing .rds files, HDF5Array summarized experiments, and banquets.}
}
\description{
Turn a directory into an environment of lazy-loading objects
}
