% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{peaks_weitrices}
\alias{peaks_weitrices}
\title{Shifts and proportions weitrices from peak counts}
\usage{
peaks_weitrices(peaks_se, n_per_row = 100, n_per_col = 100,
  remove_mispriming = TRUE, utr_or_extension_only = TRUE)
}
\description{
Convert a SummarizedExperiment of peak counts into a weitrix of shifts and a weitrix of proportions.
}
\details{
Peaks with less than min_present non-zero counts are discarded.

Only genes with two or more peaks are included.

Cells with less than min_present non-missing shifts are discarded.
}
