% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{read_polyA_peak_file_gtf}
\alias{read_polyA_peak_file_gtf}
\title{Read in the polyA peaks gtf file}
\usage{
read_polyA_peak_file_gtf(gtf_file, sep_genes = FALSE)
}
\arguments{
\item{sep_genes}{The gtf from polyApipe.py can have polyA peaks assigned 
from multiple genes (gene assigment derived from original bam intput files).
When FALSE (the default) - each peak will be present once in the table, 
and the 'peakgene' can have multiple comma separated values. 
If set to TRUE, a single peak around multiple genes will have duplicate 
rows - one per gene, and each peakgene column will contain only one gene.
(Default=FALSE)}

\item{peak_info_file}{GTF formatted peak file _specfically_ as output from polyApipe.py}
}
\description{
The polyApipe.py script will write a gtf formatted file with the polyA peak 
locations. This function will read it into a tabular format in R. 
This is not a general gff/gtf reader funtion -  
it will break on probably any other gff!
}
\examples{

peak_info_file <- system.file("extdata", "demo_dataset/demo_polyA_peaks.gff", package = "polyApiper") 
peak_info_table <- read_polyA_peak_file_gtf(peak_info_file)


}
